# gzip: pool gzip.Writer to reduce allocation