```
benchmark           old ns/op     new ns/op     delta
BenchmarkGzip-4     2219756       2123427       -4.34%

benchmark           old allocs     new allocs     delta
BenchmarkGzip-4     25             8              -68.00%

benchmark           old bytes     new bytes     delta
BenchmarkGzip-4     959566        146507        -84.73%
```