@mholt 2017-04-26
Great work @tw4452852 :) I think this will help with #735 -- probably enough that I will close that (unless they can show it's still a significant problem and is the same problem).